FIxed compiler warnings in printf conversions.

+++ src/bacnet/basic/sys/lighting_command.c
@@ -37,11 +37,14 @@ static void lighting_command_tracking_value_notify(
             }
             data->Tracking_Value_Callback(data->Key, old_value, value);
         } else {
-            debug_printf("Lighting-Command[%lu]-Out-of-Service\n", data->Key);
+            debug_printf(
+                "Lighting-Command[%lu]-Out-of-Service\n",
+                (unsigned long)data->Key);
         }
     } else {
         debug_printf(
-            "Lighting-Command[%lu]-Tracking-Value=%f\n", data->Key, value);
+            "Lighting-Command[%lu]-Tracking-Value=%f\n",
+            (unsigned long)data->Key, (double)value);
     }
 }
